third_party/WebKit/LayoutTests/fast/dom/Window/showModalDialog-invalid-arguments.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Checks that passing invalid strings to window.showModalDialog() is properly handled.");

if (window.testRunner)
    testRunner.setCanOpenWindows();

var nonConvertibleToString = { toString: function() { throw "Exception in toString()"; } };
var newWindow = null;
shouldThrow("newWindow = window.showModalDialog(nonConvertibleToString)", "'Exception in toString()'");
shouldBeNull("newWindow");
shouldThrow("newWindow = window.showModalDialog('about:blank', '', nonConvertibleToString)", "'Exception in toString()'");
shouldBeNull("newWindow");
</script>
</body>
</html>
