third_party/WebKit/LayoutTests/fast/dom/Window/showModalDialog-invalid-arguments-expected.txt

Checks that passing invalid strings to window.showModalDialog() is properly handled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS newWindow = window.showModalDialog(nonConvertibleToString) threw exception Exception in toString().
PASS newWindow is null
PASS newWindow = window.showModalDialog('about:blank', '', nonConvertibleToString) threw exception Exception in toString().
PASS newWindow is null
PASS successfullyParsed is true

TEST COMPLETE
